backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.36
google-genai>=1.0.0
tenacity==8.2.3
python-dotenv==1.0.0
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
